cluster group matches via home timezone dates
for any timezone/locale, ensure that the group stage matches from the same day in host timezone appear on the same calendar day. this is to avoid days with early AM games and late PM games.

for a group of matches on the same day in the host timezone, we'll show them on the same day in the page's timezone. sometimes this will mean that matches will be shown with +24 hour times, to indicate "the following morning".